Refactor: api 환경변수 적용
