+++ client/src/api/fetchAPI.ts
@@ -41,7 +41,7 @@ export const postReservation = async (
   body: BodyType
 ) => {
   const data = await axios
-    .post(`${SERVER_HOST}/reservation`, JSON.stringify(body), {
+    .post(`${SERVER_HOST}/reservation/${performanceId}`, JSON.stringify(body), {
       headers: {
         'Content-Type': 'application/json',
         Authorization: getCookie('accessToken'),
