+++ client/src/pages/sign/SignUp.tsx
@@ -14,6 +14,8 @@ import {
 import { useNavigate } from 'react-router-dom';
 import { H1Title } from '../../utils/SlideUp';
 
+const SERVER_HOST = process.env.REACT_APP_SERVER_HOST;
+
 interface IForm {
   email: string;
   password: string;
@@ -91,7 +93,7 @@ const SignUpPage = () => {
   /** 회원정보를 서버로 전송하는 ajax 함수 */
   const usePostSignUp = (data: IForm) => {
     axios
-      .post('https://103f-121-187-22-182.ngrok-free.app/member', data, {
+      .post(`${SERVER_HOST}/member`, data, {
         headers: { 'Content-Type': 'application/json' },
       })
       .then(response => {
@@ -108,10 +110,7 @@ const SignUpPage = () => {
   /** 이메일 중복검사하는 ajax 함수 */
   const useGetDuplicateEmail = (emailData: string | null) => {
     axios
-      .post(
-        'https://103f-121-187-22-182.ngrok-free.app/member/duplicate/email',
-        { email: emailData }
-      )
+      .post(`${SERVER_HOST}/member/duplicate/email`, { email: emailData })
       .then(response => {
         if (response.status === 200) {
           // 사용 가능한 이메일일 경우
@@ -136,10 +135,9 @@ const SignUpPage = () => {
   /** 닉네임 중복검사하는 ajax 함수 */
   const useGetDuplicateNickname = (nicknameData: string | null) => {
     axios
-      .post(
-        'https://103f-121-187-22-182.ngrok-free.app/member/duplicate/nickname',
-        { nickname: nicknameData }
-      )
+      .post(`${SERVER_HOST}/member/duplicate/nickname`, {
+        nickname: nicknameData,
+      })
       .then(response => {
         if (response.status === 200) {
           // 사용 가능한 이메일일 경우
